pyproject.toml
[tool.poetry]
name = "mai-ml-cp-backend"
version = "0.1.0"
description = ""
authors = ["Kirill Efimenko <vivichv9@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.5"
databases = "^0.9.0"
asyncpg = "^0.30.0"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
python-dotenv = "^1.0.1"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
